summariser/Summariser/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-speech>=2.22.0
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.10.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-magic>=0.4.27
pymongo>=4.5.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.25.0
PyPDF2>=3.0.1
pydub>=0.25.1
psutil>=5.9.0
google-generativeai>=0.5.0
google-cloud-texttospeech>=2.29.0

# Development and testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
reportlab>=4.0.4
